+++ src/Search.php
@@ -5,7 +5,7 @@
 
 namespace Geniem\RediPress;
 
-use Geniem\RediPress\Admin,
+use Geniem\RediPress\Settings,
     Geniem\RediPress\Redis\Client,
     Geniem\RediPress\Utility;
 use function GuzzleHttp\Promise\each;
@@ -69,11 +69,12 @@ class Search {
      * @param array  $index_info Index information.
      */
     public function __construct( Client $client, array $index_info ) {
+        $settings = new Settings();
         $this->client     = $client;
         $this->index_info = $index_info;
 
         // Get the index name from settings
-        $this->index = Admin::get( 'index' );
+        $this->index = $settings->get( 'index' );
 
         // Add search filters
         add_filter( 'posts_pre_query', [ $this, 'posts_pre_query' ], 10, 2 );
@@ -145,6 +146,10 @@ class Search {
                 'FT.SEARCH',
                 $command
             );
+
+            if ( ! is_array( $results ) ) {
+                $results = [];
+            }
 
             $index = 0;